On branch search_screen created separate folders for search_screen, changed path in others files to search_screen.dart 	modified:   lib/generated/intl/messages_en.dart 	modified:   lib/generated/l10n.dart 	modified:   lib/l10n/intl_en.arb 	modified:   lib/screens/home_screens/home_screen.dart 	renamed:    lib/screens/search_screen.dart -> lib/screens/search_screens/search_screen.dart 	modified:   lib/utils/theme.dart 	modified:   lib/widgets/sliver_appbar.dart
