+++ lib/widgets/sliver_appbar.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:ionicons/ionicons.dart';
 import 'package:project_lagann/screens/notification_screens/notifications_screen.dart';
-import '../screens/search_screen.dart';
+import '../screens/search_screens/search_screen.dart';
 import '../utils/constants.dart';
 
 class SlivAppBar extends StatelessWidget {
